requirements.txt
requests
python-dotenv
feedparser
google-generativeai